pyproject.toml
[project]
name = "dataviz-pro"
version = "0.1.0"
description = "Lightweight Streamlit app for CSV/Excel analysis with Plotly; optional AI, database, forecasting, and API extras."
requires-python = ">=3.11"
readme = "README.md"
license = { text = "MIT" }
dependencies = [
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "openpyxl>=3.1.5",
    "plotly>=6.1.2",
    "kaleido>=0.2.1",
    "scikit-learn>=1.7.0",
    "scipy>=1.15.3",
    "statsmodels>=0.14.0",
    "streamlit>=1.45.1",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "geopy>=2.4.0",
    "openai>=1.88.0",
    "groq>=0.9.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
database = [
    "pymongo>=4.6.0",
    "dnspython>=2.0.0",
    "psycopg2-binary>=2.9.0",
]
forecasting = [
    "prophet>=1.1.5",
]
api = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
]
reports = [
    "schedule>=1.2.0",
    "APScheduler>=3.10.0",
]
all = [
    "pymongo>=4.6.0",
    "dnspython>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "prophet>=1.1.5",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
    "schedule>=1.2.0",
    "APScheduler>=3.10.0",
]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.27.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = [
    "ab_testing",
    "api_server",
    "app",
    "data_context",
    "auto_viz",
    "forecasting_engine",
    "formula_builder",
    "report_scheduler",
    "start_fast",
    "streaming_engine",
    "test_db",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-q"
